// src/app/shared/models/Car.ts
// TODO: Create car model for project
export class Car {
	id: string;
	make: string;
	year: string;
	color: string;
	price: string;
	priceValue: string;
	hasSunroof: string;
	isFourWheelDrive: string;
	hasLowMiles: string;
	hasPowerWindows: string;
	hasNavigation: string;
	hasHeatedSeats: string;
}